Fix tackpad logic with one check and all check

+++ public/js/tackpad.js
@@ -11,11 +11,42 @@ function checkAllOffeneTasks(source) {
     for (var i = 0, n = checkboxes.length; i < n; i++) {
         checkboxes[i].checked = source.checked;
     }
-
-    if (document.getElementById("checkAllOffeneTasks").checked) {
+    elements_offen = document.getElementsByClassName("offene_tasks");
+    for (var i = 0; i < elements_offen.length; i++) {
+        if (elements_offen[i].checked == true) {
+            counter_offen++;
+        }
+    }
+    if (document.getElementById("checkAllOffeneTasks").checked == true && counter_offen == 1){
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
+    } else if (document.getElementById("checkAllOffeneTasks").checked == true && counter_offen > 1){
         document.getElementById("deleteAllOffeneTasks").style.display = "block";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
+    } else if (document.getElementById("checkAllOffeneTasks").checked == false && counter_offen > 1){
+        document.getElementById("deleteAllOffeneTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
+    }  
+    else if (document.getElementById("checkAllOffeneTasks").checked == false){
+        document.getElementById("deleteAllOffeneTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
     } else {
         document.getElementById("deleteAllOffeneTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "inline-block";
+        document.getElementById("loeschen").style.display = "inline-block";
+        document.getElementById("freigeben").style.display = "inline-block";
+        document.getElementById("erledigt").style.display = "inline-block";
     }
 }
 
@@ -24,11 +55,41 @@ function checkAllErledigteTasks(source) {
     for (var i = 0, n = checkboxes.length; i < n; i++) {
         checkboxes[i].checked = source.checked;
     }
-
-    if (document.getElementById("checkAllErledigteTasks").checked) {
+    elements_erledigt = document.getElementsByClassName("erledigte_tasks");
+    for (var i = 0; i < elements_erledigt.length; i++) {
+        if (elements_erledigt[i].checked == true) {
+            counter_erledigt++;
+        }
+    }
+    if (document.getElementById("checkAllErledigteTasks").checked == true && counter_offen == 1){
+        document.getElementById("bearbeiten").style.display = "inline-block";
+        document.getElementById("loeschen").style.display = "inline-block";
+        document.getElementById("freigeben").style.display = "inline-block";
+        document.getElementById("erledigt").style.display = "inline-block";
+    } else if (document.getElementById("checkAllErledigteTasks").checked == true && counter_erledigt > 1){
         document.getElementById("deleteAllErledigteTasks").style.display = "block";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
+    } else if (document.getElementById("checkAllErledigteTasks").checked == false && counter_erledigt > 1){
+        document.getElementById("deleteAllErledigteTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
+    } else if (document.getElementById("checkAllErledigteTasks").checked == false){
+        document.getElementById("deleteAllErledigteTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
+        document.getElementById("freigeben").style.display = "none";
+        document.getElementById("erledigt").style.display = "none";
     } else {
         document.getElementById("deleteAllErledigteTasks").style.display = "none";
+        document.getElementById("bearbeiten").style.display = "inline-block";
+        document.getElementById("loeschen").style.display = "inline-block";
+        document.getElementById("freigeben").style.display = "inline-block";
+        document.getElementById("erledigt").style.display = "inline-block";
     }
 }
 
@@ -53,12 +114,20 @@ function getId_for_offen(id) {
             counter_offen++;
         }
     }
-    if (elements_offen.length == counter_offen) {
+    
+    if (elements_offen.length == counter_offen && counter_offen == 1) {
+        document.getElementById("checkAllOffeneTasks").checked = true;
+        document.getElementById("bearbeiten").style.display = "inline-block";
+        document.getElementById("loeschen").style.display = "inline-block";
+        document.getElementById("freigeben").style.display = "inline-block";
+        document.getElementById("erledigt").style.display = "inline-block";
+    } else if (elements_offen.length == counter_offen && counter_offen != 1) {
         document.getElementById("checkAllOffeneTasks").checked = true;
         document.getElementById("deleteAllOffeneTasks").style.display = "block";
         document.getElementById("bearbeiten").style.display = "none";
         document.getElementById("loeschen").style.display = "none";
     } else if (counter_offen == 1) {
+        document.getElementById("checkAllOffeneTasks").checked = false;
         document.getElementById("deleteAllOffeneTasks").style.display = "none";
         document.getElementById("bearbeiten").style.display = "inline-block";
         document.getElementById("loeschen").style.display = "inline-block";
@@ -72,6 +141,7 @@ function getId_for_offen(id) {
         document.getElementById("freigeben").style.display = "none";
         document.getElementById("loeschen").style.display = "inline-block";
     } else if (counter_offen == 0) {
+        document.getElementById("checkAllOffeneTasks").checked = false;
         document.getElementById("deleteAllOffeneTasks").style.display = "none";
         document.getElementById("bearbeiten").style.display = "none";
         document.getElementById("loeschen").style.display = "none";
@@ -97,16 +167,19 @@ function getId_for_erledigt(id) {
     for (var i = 0; i < elements_erledigt.length; i++) {
         if (elements_erledigt[i].checked == true) {
             counter_erledigt++;
-            console.log("Eins erledigt!");
         }
     }
-    if (elements_erledigt.length == counter_erledigt) {
-      document.getElementById("checkAllErledigteTasks").checked = true;
-      document.getElementById("deleteAllErledigteTasks").style.display = "block";
-      document.getElementById("bearbeiten").style.display = "none";
-      document.getElementById("loeschen").style.display = "none";
-      document.getElementById("freigeben").style.display = "none";
-      document.getElementById("erledigt").style.display = "none";
+    if (elements_erledigt.length == counter_erledigt && counter_erledigt == 1) {
+        document.getElementById("checkAllErledigteTasks").checked = true;
+        document.getElementById("bearbeiten").style.display = "inline-block";
+        document.getElementById("loeschen").style.display = "inline-block";
+        document.getElementById("freigeben").style.display = "inline-block";
+        document.getElementById("erledigt").style.display = "inline-block";
+    } else if (elements_erledigt.length == counter_erledigt && counter_erledigt != 1) {
+        document.getElementById("checkAllErledigteTasks").checked = true;
+        document.getElementById("deleteAllErledigteTasks").style.display = "block";
+        document.getElementById("bearbeiten").style.display = "none";
+        document.getElementById("loeschen").style.display = "none";
     } else if (counter_erledigt == 1) {
       document.getElementById("checkAllErledigteTasks").checked = false;
       document.getElementById("deleteAllErledigteTasks").style.display = "none";
@@ -122,6 +195,7 @@ function getId_for_erledigt(id) {
       document.getElementById("freigeben").style.display = "none";
       document.getElementById("loeschen").style.display = "inline-block";
     } else if (counter_erledigt == 0) {
+      document.getElementById("checkAllErledigteTasks").checked = false;
       document.getElementById("deleteAllErledigteTasks").style.display = "none";
       document.getElementById("bearbeiten").style.display = "none";
       document.getElementById("loeschen").style.display = "none";
